move stories and style files into subfolders
