+++ client/.storybook/main.ts
@@ -10,7 +10,7 @@ function getAbsolutePath(value: string): any {
 	return dirname(require.resolve(join(value, "package.json")));
 }
 const config: StorybookConfig = {
-	stories: ["../src/**/*.mdx", "../src/**/*.stories.@(js|jsx|mjs|ts|tsx)"],
+	stories: ["../src/**/*.mdx", "../src/**/stories/*.stories.@(js|jsx|mjs|ts|tsx)"],
 	addons: [
 		getAbsolutePath("@storybook/addon-essentials"),
 		getAbsolutePath("@chromatic-com/storybook"),
